Developping an interface to openTDSB
The idea is to store the time series in a openTSDB instance to plot them
easily with Grafana. This does not replace the clone db, as openTSDB is
limited to time series and has a different structure.

Elements in place:
- the fork of a basic otsdb client
- the initial version of a class to transfer eeDomus data to openTSDB

Missing:
- the client is incomplete
- the utility class is untested
- nothing is integrated with the existing workflow

Target: a small utility to do batch transfer of past data + an option to
copy data at the time of the sync operation with eeDomus.